\name{mirageTract-methods}
\docType{methods}
\alias{mirageTract-methods}
\alias{mirageTract,Tract-method}

\title{ Methods for \code{mirageTract} object construction in Package
  \pkg{sampSurf} }

\description{
There is only one existing method available for the
  \code{\link{mirageTract}} generic function, which will generate valid
  objects of class \code{"\linkS4class{mirageTract}"}.
}

\section{Methods}{
\describe{

\item{\strong{\code{signature(tract = "Tract")}}}{ }

  \item{}{\bold{usage\ldots}
    \preformatted{mirageTract(tract,
            ... )}
	 
	\itemize{
	 
	  \item{\code{tract}:}{ A valid tract object.}
	 
	  \item{\code{\dots}:}{ Gobbled.}
	  
	} %itemize
  } %item


}} %describe, section

\keyword{methods}
%\keyword{ ~~ other possible keyword(s) ~~ }
